Change examples to use WebApplication V1
