+++ examples/Blazor/Server/Program.cs
@@ -16,23 +16,35 @@
 
 #endregion
 
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.ResponseCompression;
+using Server.Services;
 
-namespace Server
+var builder = WebApplication.CreateBuilder(args);
+builder.Services.AddGrpc();
+builder.Services.AddResponseCompression(opts =>
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+    opts.MimeTypes = ResponseCompressionDefaults.MimeTypes.Concat(
+        new[] { "application/octet-stream" });
+});
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                });
-    }
+var app = builder.Build();
+app.UseResponseCompression();
+
+if (builder.Environment.IsDevelopment())
+{
+    app.UseDeveloperExceptionPage();
+    app.UseWebAssemblyDebugging();
 }
+
+app.UseStaticFiles();
+app.UseBlazorFrameworkFiles();
+
+app.UseRouting();
+
+app.UseGrpcWeb();
+
+app.MapGrpcService<WeatherService>().EnableGrpcWeb();
+app.MapGrpcService<CounterService>().EnableGrpcWeb();
+app.MapFallbackToFile("index.html");
+
+app.Run();
