+++ examples/Blazor/Server/Server.csproj
@@ -2,6 +2,7 @@
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
+    <UserSecretsId>46982f83-f153-443e-b589-4b2bc7b5945e</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
