+++ examples/Browser/Server/wwwroot/package.json
@@ -5,7 +5,6 @@
   "devDependencies": {
     "@grpc/proto-loader": "^0.3.0",
     "google-protobuf": "^3.6.1",
-    "grpc": "^1.15.0",
     "grpc-web": "^1.0.0",
     "webpack": "^4.16.5",
     "webpack-cli": "^3.1.0"
